+++ frontend/src/components/Chart/index.js
@@ -1,5 +1,4 @@
 import React, { useRef, useLayoutEffect } from 'react';
-import './App.css';
 import * as am4core from "@amcharts/amcharts4/core";
 import * as am4charts from "@amcharts/amcharts4/charts";
 import am4themes_animated from "@amcharts/amcharts4/themes/animated";
