+++ frontend/src/css/StockManagement.css
